// src/interfaces/IDatabase.ts
import { IAction } from "./IAction";
import { IDatabaseLocation } from "./IDatabaseLocation";
import { IDatabaseType } from "./IDatabaseType";
import { IForm } from "./IForm";
import { ILicenceType } from "./ILicenceType";
import { IOrganization } from "./IOrganization";
import { ISector } from "./ISector";
import { ITable } from "./ITable";

export type IDatabase = {
  id: number;
  name: string;
  org_id: number;
  description: string;
  spec: [{ id: string }];
  spec_other: string;
  db_type: number;
  db_type_other: string;
  db_location: number;
  db_location_other: string;
  sector: string;
  sector_other: string;
  table_count: number;
  start_date: number;
  licence_type: string;
  licence_type_other: string;
  opendata_url: string;
  is_form: true;
  is_active: false;
  is_integrated: false;
  created_date: string;
  forms: IForm[];
  tables: ITable[];
  organization: IOrganization;
  databaseType: IDatabaseType;
  databaseLocation: IDatabaseLocation;
  sectors: ISector;
  licenceType: ILicenceType;
  version: string;
  createdUser: number;
  actions: IAction;
};